src/geminiservice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-generativeai==0.3.0
google-cloud-firestore==2.13.0
google-cloud-pubsub==2.18.0
redis==5.0.1
python-multipart==0.0.6
httpx==0.25.0
